+++ lib/yt/core.rb
@@ -9,7 +9,6 @@ require 'yt/resource'
 require 'yt/response'
 
 require 'yt/channel'
-require 'yt/group'
 require 'yt/playlist'
 require 'yt/playlist_item'
 require 'yt/video'